+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+### Added
+- Add sunburst plots.
 ### Fixed
 - `SeqSummary` read quality histograms only showing one bar which is too thin to see when all reads have the same quality score.